+++ stDrosophila/__init__.py
@@ -1,9 +1,7 @@
 """
 Toolkit for analyzing Drosophila spatial transcriptome data.
 """
-from . import envs
 from . import io
-from . import od
 from . import pl
 from . import pp
 from . import tl
