--- stDrosophila/obtain_dataset/__init__.py
@@ -1,3 +1,0 @@
-from .from_flymine import fm_BDGP, fm_gene2GO
-from .from_flybase import *
-from .from_flyaltas2 import flyatlas2
