--- stDrosophila/od.py
@@ -1,2 +1,0 @@
-""""""
-from .obtain_dataset import *
